Split stream and release packages
The goal of this project is for API clients to query streams, not
releases directly as those are more of an implementation detail
today.

Split them into separate packages so that using `/streams` will
get just that.

This does require duplicating the common types.

And also golint really wants us to now rename e.g.
`release.ReleaseArch` to just `release.Arch`, but
will do that as a followup.