stream: Add helper APIs for AWS
In updating openshift-install to use this I really wanted to have
nice error prefixing if a particular architecture/region wasn't
found.  These new APIs handle that and also are what basically
everyone interacting with AWS will want.

(I may update the example to use them in the future, but it's
 also useful to show traversing the "raw" data structures)