Add an uncompressed sha256 to release and streams
Moving this from
https://github.com/coreos/coreos-assembler/pull/2000#issuecomment-763152177

I'm trying to port openshift-install to streams, and it currently
is using the uncompressed sha256 - and further it's exposed in
some cases to users.

Since compression isn't always easily reproducible, I think what
cosa is doing in having both checksums is a good thing; might
as well do belt-and-suspenders.

To emphasize we already have this data from the cosa
metadata, this is just adding support for it to the stream
and release metadata.