Add RHELCoreOSExtensions with Azure disk URL
This contains just a URL for now because that's all
the current RHCOS cosa metadata has.  I'm trying
to add e.g. the Azure Blob storage md5 information as
well as the full size+sha256, but in practice this
data is just the uncompressed VHD; anyone who wants
to do "offline" verification outside of Azure can
replicate that.

For FCOS we may end up uploading the image too,
though we hope there to end up in the Marketplace.

For now, let's stick this off an explicit extension area.

Closes: https://github.com/coreos/stream-metadata-go/issues/13