artifact: Add a method to download+validate checksum
Ideally of course we'd validate the GPG signatures too but
that runs into the problem of us shipping the keys in this
library or depending on them externally.

I'd also like to add the size to the stream metadata, because
it allows e.g. detecting the case when a server is feeding us
too much data (or too little).

Will circle back to that later.