stream: make Artifact.Signature omitempty
RHCOS artifacts are not signed.  If we omit signature URLs from the stream
metadata, we currently get a field with an empty string as a value.  Since
artifacts will likely still be unsigned for a while, acquiesce to reality
and correctly omit the field.

xref https://github.com/openshift/installer/pull/5268