Add support for IBMCloud/PowerVS images
Today we generate metadata for PowerVS in RHCOS, and we would need this to be updated in the installer in the near future, hence this change. IBMCloud and PowerVS have the same format although as of today there are no IBMCloud objects being generated. The naming here is to use "object" instead of image since these are not traditional bootimages like in other clouds.

FCOS is still waiting on ppc64le hardware to have P images generated so could not add any tests related to this.