Add digest reference to KubeVirt image description
In the current design, the KubeVirt image is presumed to be a fully-
qualified reference, including a container digest.  But that implies that
the user should reference the image by digest, when we'd prefer that they
reference it by stream-specific tag.

Define the existing "image" field to contain the preferred form of
fully-qualified reference for the image, which might involve either a
tag or a digest.  We should still record the unique identifier of the
current image (as we do for GCP images), so add a "digest-ref" field
which always contains a fully-qualified reference with digest.